refactored encoding/decoding to produce DecodedExchange namedtuples that contain the response name, added desired client code and a test for the BaseClient

+++ hardsync/example/contract.py
@@ -1,6 +1,6 @@
 class MeasureVoltage:
     class Request:
-        integration_time: int
+        integration_time: float
         channel: int
 
     class Response:
